django_hh_project/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.10
django-allauth==0.39.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
GoogleAppEngineCloudStorageClient==1.9.22.1
idna==2.8
oauthlib==3.1.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
